Door glitch fixed and hierarchy Cleanup
The doors had a exploit that let the player open up areas they shouldnt be yet by putting a candle through the door.
There is a a tag check so only the player and enemy can trigger a door.

I also attempted to clean the hierarchy so that in the future adding additional features will be easier.

and minor clipping texture where fixed as well